更正 資料型態對應

+++ src/main/java/idv/tia201/g2/web/chat/vo/ChatSessions.java
@@ -22,8 +22,8 @@ public class ChatSessions {
     @Column(name = "chat_status_id")
     private Integer chatStatusId;
     @Column(name = "attender_id")
-    private Integer attenderId;
+    private Long attenderId;
     @Column(name = "administrator_id")
-    private Integer administratorId;
+    private Long administratorId;
 
 }
